refactor: split bundle.py/cli_options.py/cli_compare_release.py (G38 Phase 15 file-split prereq)
Continues the G38 Phase 15 file-split prerequisite started in PR #971
(buildsource/inline.py split): splits the three remaining candidate files
the plan named as blocking new .abicheck.yml `bundle_variants:` CLI/config
surface.

- bundle.py (1999 -> 873 lines): detection helpers and heuristic
  primitives moved to new siblings bundle_detectors.py and
  bundle_detector_heuristics.py (two files, since one combined module
  would have exceeded the 800-line cap for a new file).
- cli_options.py (1977 -> 1504 lines): the release-fanout/build-source/
  header-graph/evidence option-group cluster moved to
  frontends/cli/options/release.py, matching the existing
  contract.py/profiles.py/secondary_output.py sibling pattern.
- cli_compare_release.py (1995 -> 820 lines, pinned at its exact
  adoption-time line count per ADR-061 after an earlier attempt here grew
  two frozen files): the per-pair/per-library comparison engine and the
  matrix-result/output/gating engine moved to two new siblings,
  cli_compare_release_pairwise.py and cli_compare_release_matrix.py, each
  importing directly from cli_compare_release_helpers.py so neither
  pinned file grows.

Each moved module re-exports every name an existing caller or test
imports directly, for back-compat. Several tests that monkeypatched a
cli_compare_release private helper by its old module path were updated
to patch the function's new home instead, since the helper's own callers
now resolve it through that module's namespace, not the old one.

architecture/modules.yaml and check_ai_readiness.py's IMPORT_CYCLE_ALLOWLIST
updated to register the six new files, following the established
"split of an already-member module carries no new dependency direction"
precedent already documented there for prior splits.
