+++ abicheck/bundle.py
@@ -55,6 +55,13 @@ Bundle findings use the ``ChangeKind.BUNDLE_*`` values registered in
 :mod:`abicheck.change_registry`. They participate in policy classification,
 suppression, severity, and reporter machinery identically to per-library
 ``Change`` entries.
+
+Individual finding-producers (the ``_detect_*`` functions this module's
+``compare_bundle``/``audit_bundle`` orchestrate) and the heuristic
+primitives they share live in :mod:`abicheck.bundle_detectors` and its own
+sibling :mod:`abicheck.bundle_detector_heuristics`, split out purely to
+stay under the AI-readiness 2000-line hard cap -- see either module's own
+docstring.
 """
 
 from __future__ import annotations
@@ -65,6 +72,25 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import TYPE_CHECKING
 
+from .bundle_detector_heuristics import (  # noqa: F401  (re-exported for back-compat)
+    _build_demangled_index as _build_demangled_index,
+    _detect_manifest_drift,
+    _detect_soname_skew,
+    _looks_system_symbol as _looks_system_symbol,
+    _match_entry as _match_entry,
+    _path_looks_like_elf,
+    _soname_skew_findings as _soname_skew_findings,
+    _strip_namespace_prefix as _strip_namespace_prefix,
+)
+from .bundle_detectors import (
+    _detect_intra_dep_removed,
+    _detect_intra_dep_signature_changed,
+    _detect_intra_type_changed,
+    _detect_library_structural_changes,
+    _detect_provider_changed,
+    _detect_unresolved_intra_dependency,
+    _detect_version_drift,
+)
 from .bundle_manifest import (  # noqa: F401  (re-exported for back-compat)
     InstantiationManifest as InstantiationManifest,
     ManifestEntry as ManifestEntry,
@@ -73,7 +99,6 @@ from .bundle_manifest import (  # noqa: F401  (re-exported for back-compat)
 )
 from .bundle_models import (  # noqa: F401  (re-exported for back-compat)
     DEFAULT_SYSTEM_PROVIDERS as DEFAULT_SYSTEM_PROVIDERS,
-    PROMOTABLE_C_BOUNDARY_SIGNATURE_BREAK_KINDS,
     BundleDiffResult as BundleDiffResult,
     BundleFinding as BundleFinding,
     BundleSnapshot as BundleSnapshot,
@@ -81,24 +106,16 @@ from .bundle_models import (  # noqa: F401  (re-exported for back-compat)
     ProviderEntry as ProviderEntry,
     ResolutionGraph as ResolutionGraph,
     basename_to_bundle_key,
-    consumer_resolves_via_provider as _consumer_resolves_via_provider,
-    diff_change_is_breaking as _diff_change_is_breaking,
 )
-from .bundle_resolution_reachability import (
-    cached_reachable_intra_libraries as _cached_reachable_intra_libraries,
-    reachable_intra_libraries as _reachable_intra_libraries,
-)
-from .bundle_soname import hard_link_alias_basenames, soname_matches_providers
-from .checker_policy import ChangeKind, Verdict, compute_verdict, policy_kind_sets
+from .bundle_soname import hard_link_alias_basenames
+from .checker_policy import Verdict, compute_verdict
 from .checker_types import DiffResult
-from .elf_metadata import ElfMetadata, ElfSymbol, SymbolBinding, parse_elf_metadata
+from .elf_metadata import ElfMetadata, parse_elf_metadata
 
 if TYPE_CHECKING:
-    from .diff_cpp_patterns import BundleMember
     from .policy_file import PolicyFile
 
 log = logging.getLogger(__name__)
-
 
 # ---------------------------------------------------------------------------
 # Bundle snapshot construction
@@ -264,7 +281,10 @@ def build_bundle_snapshot_from_metadata(
     for name, meta in metadata.items():
         deadline.check()
         if meta is None or (
-            not meta.soname and not meta.symbols and not meta.imports and not meta.needed
+            not meta.soname
+            and not meta.symbols
+            and not meta.imports
+            and not meta.needed
         ):
             log.debug("bundle: skipping empty metadata for %s", name)
             continue
@@ -855,1145 +875,3 @@ def compare_bundle(
         policy=policy,
         policy_file=policy_file,
     )
-
-
-# ---------------------------------------------------------------------------
-# Detection helpers
-# ---------------------------------------------------------------------------
-
-
-def _detect_library_structural_changes(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-) -> list[BundleFinding]:
-    """Detect libraries that appeared or disappeared.
-
-    Only emits :class:`ChangeKind.BUNDLE_LIBRARY_REMOVED` when the missing
-    library exported at least one symbol consumed by a surviving sibling
-    in the old bundle — that is, the removal actually breaks the bundle's
-    internal contract. A removal that broke nothing internally is handled
-    by the existing ``--fail-on-removed-library`` flow.
-    """
-    findings: list[BundleFinding] = []
-    old_names = set(old.libraries)
-    new_names = set(new.libraries)
-
-    for added in sorted(new_names - old_names):
-        findings.append(
-            BundleFinding(
-                kind=ChangeKind.BUNDLE_LIBRARY_ADDED,
-                symbol=added,
-                description=f"New library {added} appears in the bundle.",
-                provider_library=added,
-            ),
-        )
-
-    for removed in sorted(old_names - new_names):
-        # Was the removed lib actually depended on by a surviving sibling?
-        # Only emit a bundle finding when the removal actually breaks the
-        # internal contract. Stand-alone library removal is handled by
-        # the existing --fail-on-removed-library flow.
-        old_meta = old.metadata.get(removed)
-        consumers: list[str] = []
-        if old_meta is not None:
-            exports = {s.name for s in old_meta.symbols}
-            for sib_name, sib_meta in old.metadata.items():
-                if sib_name == removed or sib_name not in new.metadata:
-                    continue
-                if any(imp.name in exports for imp in sib_meta.imports):
-                    consumers.append(sib_name)
-        if not consumers:
-            continue
-        findings.append(
-            BundleFinding(
-                kind=ChangeKind.BUNDLE_LIBRARY_REMOVED,
-                symbol=removed,
-                description=(
-                    f"Library {removed} removed from the bundle; "
-                    f"depended on by: {', '.join(sorted(consumers))}"
-                ),
-                provider_library=removed,
-                affected_libraries=consumers,
-            ),
-        )
-
-    return findings
-
-
-def _detect_intra_dep_removed(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-    system_providers: set[str],
-    explicit_providers: set[str],
-) -> list[BundleFinding]:
-    """Find imports in the new bundle that no sibling provides.
-
-    Excludes imports satisfied by system libraries (out of bundle scope by
-    design). Classification uses provider/version evidence first
-    (:func:`_import_is_external`) and the symbol-name allow-list
-    (``DEFAULT_SYSTEM_SYMBOLS`` / ``_looks_system_*``) as a fallback.
-    Excludes weak imports (linker treats unresolved weak as 0/NULL).
-
-    A consumer's import is treated as system-provided when every DT_NEEDED
-    edge it carries that resolves *outside* the bundle is in the
-    ``system_providers`` allow-list -- but only when either (a) *this
-    consumer* never reached a *version-compatible* in-bundle sibling
-    providing this symbol in ``old`` (checked via reachability, not merely
-    a same-named provider existing somewhere in the old bundle -- an
-    unrelated consumer's own old provider must not veto a different
-    consumer's always-external dependency, nor may a provider whose
-    version could never have satisfied this consumer's own reference), or
-    (b) the user explicitly named at least one of this consumer's
-    remaining sonames via ``explicit_providers``. Without (a)/(b), a
-    sibling provider and its DT_NEEDED edge could have been dropped by the
-    same refactor that left this consumer needing libc -- so absent
-    either, the symbol-name check below still has to agree.
-    """
-    findings: list[BundleFinding] = []
-    old_reachable_cache: dict[str, set[str]] = {}
-
-    def _old_reachable(lib: str) -> set[str]:
-        if lib not in old_reachable_cache:
-            old_reachable_cache[lib] = _reachable_intra_libraries(old, lib)
-        return old_reachable_cache[lib]
-
-    for symbol, consumers in new.resolution.consumers.items():
-        providers = new.resolution.providers_for(symbol)
-        if providers:
-            continue  # someone in the bundle provides it
-        old_all_providers = old.resolution.providers_for(symbol)
-        for consumer in consumers:
-            if consumer.weak:
-                continue
-            consumer_meta = new.metadata.get(consumer.library)
-            if consumer_meta is None:
-                continue
-            # Did *this* consumer -- not merely some other consumer of the
-            # same symbol name -- previously reach a *version-compatible*
-            # old provider (docstring above; mirrors the audit-mode
-            # sibling's identical compatibility rule)?
-            if consumer.version:
-                compatible_old = {
-                    p.library
-                    for p in old_all_providers
-                    if p.version == consumer.version
-                }
-            else:
-                compatible_old = {p.library for p in old_all_providers if p.is_default}
-            ever_provided_in_bundle = bool(
-                compatible_old
-                and consumer.library in old.libraries
-                and compatible_old & _old_reachable(consumer.library)
-            )
-            if _import_is_external(consumer, consumer_meta, new):
-                continue
-            # Every non-intra DT_NEEDED on the allow-list AND (no sibling
-            # ever provided this symbol, OR the user explicitly asserted a
-            # remaining soname) -> trust it unconditionally. Otherwise fall
-            # through to the symbol-name check (docstring above).
-            # Known limitations (Codex review, shared by the audit-mode
-            # sibling below): (1) absence of a *bundle* regression is not
-            # proof of a system export (no export-table parse); (2) `all()`
-            # below is over every extra edge, not just whichever provides
-            # `symbol`. DEFAULT_SYSTEM_PROVIDERS broadened to reduce (2).
-            extra_needed = new.resolution.extra_needed.get(consumer.library, [])
-            if (
-                extra_needed
-                and all(
-                    soname_matches_providers(e, system_providers) or _looks_system(e)
-                    for e in extra_needed
-                )
-                and (
-                    not ever_provided_in_bundle
-                    or any(
-                        soname_matches_providers(e, explicit_providers)
-                        for e in extra_needed
-                    )
-                )
-            ):
-                continue
-            if symbol in DEFAULT_SYSTEM_SYMBOLS or _looks_system_symbol(symbol):
-                continue
-            findings.append(
-                BundleFinding(
-                    kind=ChangeKind.BUNDLE_INTRA_DEP_REMOVED,
-                    symbol=symbol,
-                    description=(
-                        f"{consumer.library} imports {symbol}, but no library in "
-                        f"the new bundle exports it. Runtime load of "
-                        f"{consumer.library} will fail with undefined symbol."
-                    ),
-                    consumer_library=consumer.library,
-                    affected_libraries=[consumer.library],
-                ),
-            )
-    return findings
-
-
-def _detect_unresolved_intra_dependency(
-    new: BundleSnapshot,
-    system_providers: set[str],
-) -> list[BundleFinding]:
-    """Audit-mode (no old side) sibling of :func:`_detect_intra_dep_removed`.
-
-    ADR-056 D2: ``scan --artifact-set`` has no per-library diff to read, so
-    this operates purely off the new-side resolution graph. Deliberately
-    **not** a call into :func:`_detect_intra_dep_removed`; differs in three
-    ways that matter for soundness here:
-
-    1. **Version-aware, reachability-constrained provider matching.**
-       ``providers_for(symbol)`` is name-only and set-wide.
-       ``ProviderEntry.version``/``ConsumerEntry.version`` are consulted so a
-       version mismatch (consumer needs ``foo@V2``, set only provides
-       ``foo@V1``) is not mistaken for a resolved import; when the precise
-       ``ConsumerEntry.version_soname`` is known, the match is pinned to
-       that exact provider library (GNU version *labels* are not globally
-       unique). Every candidate provider must additionally be reachable
-       from the consumer through :func:`_reachable_intra_libraries` — a
-       match on a library the consumer has no ``DT_NEEDED`` path to would
-       never actually be loaded together with the consumer.
-    2. **A narrower, explicitly-approximate suppression path for unversioned
-       imports.** Mirrors ``_detect_intra_dep_removed``'s allow-list union
-       and its non-empty guard (``extra_edges and all(...)``, never a bare
-       ``all([])``), but adds a requirement one-sided audit needs and the
-       diff-driven detector does not: the consumer must have **zero**
-       intra-bundle ``DT_NEEDED`` edges — a consumer still depending on an
-       intra-set library that simply stopped exporting the symbol has a
-       real, in-set candidate provider this coarse check cannot rule out.
-       Deliberately has **no** symbol-name-shape fallback
-       (``_looks_system_symbol``): ``--bundle-system-providers`` exists
-       specifically to cover a legitimate, non-system-shaped custom export
-       (e.g. ``vendor_init``) that a shape heuristic would never match.
-    3. Emits ``ChangeKind.BUNDLE_UNRESOLVED_INTRA_DEPENDENCY`` (not
-       ``BUNDLE_INTRA_DEP_REMOVED`` — that kind implies a diff-confirmed
-       removal, which this finding cannot claim) at
-       ``COMPATIBLE_WITH_RISK``, not ``BREAKING``: an audit has no old side
-       to confirm the symbol ever resolved.
-    """
-    findings: list[BundleFinding] = []
-    reachable_cache: dict[str, set[str]] = {}
-
-    def _reachable(lib: str) -> set[str]:
-        if lib not in reachable_cache:
-            reachable_cache[lib] = _reachable_intra_libraries(new, lib)
-        return reachable_cache[lib]
-
-    for symbol, consumers in new.resolution.consumers.items():
-        providers = new.resolution.providers_for(symbol)
-        for consumer in consumers:
-            if consumer.weak:
-                continue
-            consumer_meta = new.metadata.get(consumer.library)
-            if consumer_meta is None:
-                continue
-            reachable = _reachable(consumer.library)
-
-            if consumer.version:
-                if consumer.version_soname:
-                    # Same soname_to_name map _reachable_intra_libraries()
-                    # uses (Codex review) -- provider_library_for_soname()'s
-                    # independent heuristic has the identical
-                    # resolved-through-symlink gap, so a version_soname
-                    # naming a provider's real on-disk filename could fail
-                    # to resolve here even when that provider is genuinely
-                    # reachable.
-                    target_lib = new.resolution.soname_to_name.get(
-                        consumer.version_soname
-                    )
-                    resolved = target_lib is not None and target_lib in reachable
-                    resolved = resolved and any(
-                        p.library == target_lib and p.version == consumer.version
-                        for p in providers
-                    )
-                else:
-                    resolved = any(
-                        p.version == consumer.version and p.library in reachable
-                        for p in providers
-                    )
-            else:
-                # P2 regression (Codex review): an unversioned consumer
-                # reference can only be satisfied by an unversioned or
-                # default-version ("@@default") provider definition -- a
-                # provider that exports this symbol *only* as a non-default
-                # versioned definition ("foo@V1", not "foo@@V1") cannot
-                # satisfy it, even though the bare symbol name is reachable.
-                resolved = any(
-                    p.library in reachable and p.is_default for p in providers
-                )
-
-            if resolved:
-                continue
-            if _import_is_external(consumer, consumer_meta, new):
-                continue
-
-            if not consumer.version:
-                intra_edges = new.resolution.intra_needed.get(consumer.library, [])
-                extra_edges = new.resolution.extra_needed.get(consumer.library, [])
-                if (
-                    not intra_edges
-                    and extra_edges
-                    and all(
-                        soname_matches_providers(e, system_providers)
-                        or _looks_system(e)
-                        for e in extra_edges
-                    )
-                ):
-                    continue
-
-            findings.append(
-                BundleFinding(
-                    kind=ChangeKind.BUNDLE_UNRESOLVED_INTRA_DEPENDENCY,
-                    symbol=symbol,
-                    description=(
-                        f"{consumer.library} imports {symbol}, but no provider "
-                        "was found for it in this artifact set (audit mode — "
-                        "no old side to confirm this ever resolved)."
-                    ),
-                    consumer_library=consumer.library,
-                    affected_libraries=[consumer.library],
-                ),
-            )
-    return findings
-
-
-def _detect_intra_dep_signature_changed(
-    new: BundleSnapshot,
-    diff_by_library: dict[str, DiffResult],
-    policy: str = "strict_abi",
-) -> list[BundleFinding]:
-    """Promote provider-side signature changes to consumer-side findings.
-
-    For each confirmed *promotable* C-boundary signature break
-    (:data:`~abicheck.bundle_models.PROMOTABLE_C_BOUNDARY_SIGNATURE_BREAK_
-    KINDS`, a strict subset of ``CONFIRMED_C_BOUNDARY_SIGNATURE_BREAK_
-    KINDS``), find siblings importing the symbol that *actually resolve it
-    against this provider* (:func:`_consumer_resolves_via_provider`),
-    emitting one finding per (consumer, symbol) pair. *policy*: a
-    named-policy demotion and a ``--policy-file`` override on the diff are
-    both honored, via :func:`_diff_change_is_breaking`. Scanned including
-    ``diff.out_of_surface_changes`` (G38 Phase 14): a headerless break
-    demoted by ``--scope-public-headers`` still breaks the bundle's own
-    linkage contract (``docs/use/multi-binary.md``).
-    """
-    findings: list[BundleFinding] = []
-    seen: set[tuple[str, str, str]] = set()
-    relevant_kinds = PROMOTABLE_C_BOUNDARY_SIGNATURE_BREAK_KINDS
-    policy_sets = policy_kind_sets(policy)
-    reachable_cache: dict[str, set[str]] = {}
-
-    def _reachable(lib: str) -> set[str]:
-        return _cached_reachable_intra_libraries(new, reachable_cache, lib)
-
-    for provider_lib, diff in diff_by_library.items():
-        for change in diff.changes + diff.out_of_surface_changes:
-            if change.kind not in relevant_kinds:
-                continue
-            if not _diff_change_is_breaking(diff, change, policy_sets):
-                continue
-            consumers = new.resolution.consumers_of(change.symbol)
-            consumer_libs = sorted(
-                {
-                    c.library
-                    for c in consumers
-                    if c.library != provider_lib
-                    and _consumer_resolves_via_provider(
-                        new, c, provider_lib, change.symbol, _reachable(c.library)
-                    )
-                }
-            )
-            if not consumer_libs:
-                continue
-            for consumer_lib in consumer_libs:
-                key = (consumer_lib, provider_lib, change.symbol)
-                if key in seen:
-                    continue
-                seen.add(key)
-                findings.append(
-                    BundleFinding(
-                        kind=ChangeKind.BUNDLE_INTRA_DEP_SIGNATURE_CHANGED,
-                        symbol=change.symbol,
-                        description=(
-                            f"{consumer_lib} calls {change.symbol} (mangled name "
-                            f"unchanged) but {provider_lib} altered its DWARF "
-                            f"signature. Calling convention is now mismatched."
-                        ),
-                        consumer_library=consumer_lib,
-                        provider_library=provider_lib,
-                        old_value=change.old_value,
-                        new_value=change.new_value,
-                        affected_libraries=[consumer_lib],
-                    ),
-                )
-    return findings
-
-
-def _detect_intra_type_changed(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-    diff_by_library: dict[str, DiffResult],
-) -> list[BundleFinding]:
-    """Detect a type layout change that crosses a DSO boundary.
-
-    Conservative heuristic: a ``type_*_changed`` against type ``T`` in
-    library A counts as cross-DSO iff *some other library B* in the bundle
-    exports a symbol whose name contains ``T`` (template instantiation,
-    mangled signature reference). Catches the ``detail::``-style
-    pattern where a type defined in core leaks into algo's mangled
-    symbols. Misses extern-C function pointers that pass struct
-    references (would require type-graph propagation from DWARF, future
-    work — out of scope for ADR-023 first cut). Scanned including
-    ``diff.out_of_surface_changes`` (G38 Phase 14, rationale in
-    ``_detect_intra_dep_signature_changed``'s own docstring).
-
-    Reachability scope (ADR-027 A3 / D3.2 limitation). The public-vs-internal
-    split below is computed from ``ElfMetadata.symbols``, which is parsed from
-    ``.dynsym`` and therefore holds only **exported** (GLOBAL/WEAK,
-    non-hidden) definitions — LOCAL and hidden symbols are not retained
-    (``elf_metadata._parse_dynsym``). Consequently a consumer that references
-    the changed type **only** from its own internal/hidden functions leaves no
-    trace here: no exported symbol carries the type name, ``consumer_reach``
-    stays empty, and no finding is emitted for that consumer. The
-    ``internal_hit`` (demote-to-risk) branch therefore fires only for inputs
-    whose snapshots happen to carry local symbols (e.g. relocatable objects or
-    DWARF-derived graphs); detecting purely internal cross-DSO references from
-    a stripped shared object would require retaining non-exported symbols or a
-    full type graph, deliberately out of scope (documented per the ADR-027
-    review — the conservative miss is preferred over a parser/schema change
-    with broad blast radius).
-    """
-    findings: list[BundleFinding] = []
-    # Dedup: one finding per (consumer, provider, type) triple — multiple
-    # low-level changes (size + alignment + field-removed) against the
-    # same type would otherwise emit N copies of the same cross-DSO break.
-    seen: set[tuple[str, str, str]] = set()
-    type_kinds = {
-        ChangeKind.TYPE_SIZE_CHANGED,
-        ChangeKind.TYPE_ALIGNMENT_CHANGED,
-        ChangeKind.TYPE_FIELD_REMOVED,
-        ChangeKind.TYPE_FIELD_OFFSET_CHANGED,
-        ChangeKind.TYPE_FIELD_TYPE_CHANGED,
-        ChangeKind.TYPE_BASE_CHANGED,
-        ChangeKind.TYPE_VTABLE_CHANGED,
-        ChangeKind.INTERNAL_TYPE_LEAKS_VIA_PUBLIC_API,
-    }
-    for provider_lib, diff in diff_by_library.items():
-        for change in diff.changes + diff.out_of_surface_changes:
-            if change.kind not in type_kinds:
-                continue
-            type_name = change.symbol
-            # Look for the type name embedded in another library's symbols.
-            # ADR-027 A3/D3.2 reachability filter: classify each consumer match
-            # as reaching its *public* surface (the type leaks into a symbol the
-            # consumer itself exports — full-confidence cross-DSO break) vs only
-            # its *internal* surface (the type appears only in the consumer's
-            # local/hidden symbols — demoted to risk, not dropped, because the
-            # consumer does not re-expose the type across the boundary).
-            stripped = _strip_namespace_prefix(type_name)
-            consumer_reach: dict[str, bool] = {}  # consumer -> reaches public surface
-            for sib_name, sib_meta in new.metadata.items():
-                if sib_name == provider_lib:
-                    continue
-                public_hit = False
-                internal_hit = False
-                for sym in sib_meta.symbols:
-                    if stripped and stripped in sym.name:
-                        if _is_public_surface_symbol(sym):
-                            public_hit = True
-                            break
-                        internal_hit = True
-                if public_hit or internal_hit:
-                    consumer_reach[sib_name] = public_hit
-            for consumer_lib in sorted(consumer_reach):
-                key = (consumer_lib, provider_lib, type_name)
-                if key in seen:
-                    continue
-                seen.add(key)
-                on_public = consumer_reach[consumer_lib]
-                finding = BundleFinding(
-                    kind=ChangeKind.BUNDLE_INTRA_TYPE_CHANGED,
-                    symbol=type_name,
-                    description=(
-                        f"{provider_lib} changed type {type_name}; the type "
-                        f"is reachable from {consumer_lib}'s exported symbols. "
-                        f"{consumer_lib}'s ABI looks unchanged in isolation, "
-                        f"but every cross-DSO use of {type_name} is affected."
-                    ),
-                    consumer_library=consumer_lib,
-                    provider_library=provider_lib,
-                    affected_libraries=[consumer_lib],
-                )
-                if not on_public:
-                    # Consumer uses the type only internally → demote (disclosed,
-                    # never dropped): the cross-DSO change cannot reach the
-                    # consumer's own public ABI surface.
-                    finding.effective_verdict = Verdict.COMPATIBLE_WITH_RISK
-                    finding.modulation_reason = "consumer-internal-use"
-                    finding.modulation_rule = "bundle-reachability"
-                    finding.description = (
-                        f"{provider_lib} changed type {type_name}; {consumer_lib} "
-                        f"references it only via internal (non-exported) symbols, "
-                        f"so the change does not reach {consumer_lib}'s public ABI "
-                        f"surface — review, but not a confirmed cross-DSO break."
-                    )
-                findings.append(finding)
-    return findings
-
-
-def _is_public_surface_symbol(sym: ElfSymbol) -> bool:
-    """True when *sym* is part of a library's exported public ABI surface.
-
-    A defined GLOBAL/WEAK symbol with default/protected visibility is reachable
-    by external consumers; LOCAL binding or hidden/internal visibility is not
-    (it is the DSO's private implementation). Used by the A3 reachability filter
-    to tell a real cross-DSO break from an internal-only type reference.
-    """
-    if sym.binding not in (SymbolBinding.GLOBAL, SymbolBinding.WEAK):
-        return False
-    return sym.visibility in ("default", "protected")
-
-
-def _detect_provider_changed(
-    new: BundleSnapshot,
-    diff_by_library: dict[str, DiffResult],
-) -> list[BundleFinding]:
-    """Detect symbol provider migration within the bundle.
-
-    A symbol that was removed from library A in this release and added
-    (with the same mangled name) to library B in the same release is most
-    likely a provider move, not an ABI change. Promote both per-library
-    findings into one ``BUNDLE_PROVIDER_CHANGED`` event. Scanned including
-    ``diff.out_of_surface_changes`` (G38 Phase 14) -- deliberately still
-    with no reachability requirement, unlike its two siblings: a provider
-    move breaks an external consumer exactly as much as a sibling (ADR-023).
-    """
-    findings: list[BundleFinding] = []
-
-    removed_by: dict[str, str] = {}  # symbol -> library that removed it
-    added_by: dict[str, str] = {}  # symbol -> library that added it
-    for lib_name, diff in diff_by_library.items():
-        for change in diff.changes + diff.out_of_surface_changes:
-            if change.kind in (ChangeKind.FUNC_REMOVED, ChangeKind.VAR_REMOVED):
-                removed_by.setdefault(change.symbol, lib_name)
-            elif change.kind in (ChangeKind.FUNC_ADDED, ChangeKind.VAR_ADDED):
-                added_by.setdefault(change.symbol, lib_name)
-
-    for symbol, old_lib in removed_by.items():
-        new_lib = added_by.get(symbol)
-        if new_lib is None or new_lib == old_lib:
-            continue
-        # Confirm the symbol exists in the new bundle at the new provider.
-        providers = new.resolution.providers_for(symbol)
-        if not any(p.library == new_lib for p in providers):
-            continue
-        findings.append(
-            BundleFinding(
-                kind=ChangeKind.BUNDLE_PROVIDER_CHANGED,
-                symbol=symbol,
-                description=(
-                    f"Symbol {symbol} moved from {old_lib} to {new_lib} within "
-                    f"the bundle. Downstream consumers with DT_NEEDED on "
-                    f"{old_lib} only resolve transitively if their dependency "
-                    f"chain reaches {new_lib}."
-                ),
-                provider_library=new_lib,
-                old_value=old_lib,
-                new_value=new_lib,
-                affected_libraries=[old_lib, new_lib],
-            ),
-        )
-
-    return findings
-
-
-def _detect_version_drift(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-) -> list[BundleFinding]:
-    """Detect gnu.version_d drift on intra-bundle imports.
-
-    Compares each new-side consumer import's required version against the
-    old-side provider's defined version for the same symbol. Emits one
-    finding per symbol whose version moved.
-    """
-    findings: list[BundleFinding] = []
-
-    # Build (symbol -> old default version) from old bundle.
-    old_default_version: dict[str, str] = {}
-    for providers in old.resolution.provides.values():
-        pass  # symbols are keys
-    for sym_name, providers in old.resolution.provides.items():
-        for prov in providers:
-            if prov.version:
-                old_default_version.setdefault(sym_name, prov.version)
-                break
-
-    new_default_version: dict[str, str] = {}
-    for sym_name, providers in new.resolution.provides.items():
-        for prov in providers:
-            if prov.version:
-                new_default_version.setdefault(sym_name, prov.version)
-                break
-
-    common = set(old_default_version) & set(new_default_version)
-    for sym in sorted(common):
-        if old_default_version[sym] == new_default_version[sym]:
-            continue
-        consumers = new.resolution.consumers_of(sym)
-        consumer_libs = sorted({c.library for c in consumers})
-        if not consumer_libs:
-            continue
-        findings.append(
-            BundleFinding(
-                kind=ChangeKind.BUNDLE_INTRA_DEP_VERSION_DRIFT,
-                symbol=sym,
-                description=(
-                    f"Symbol {sym} now exported at version "
-                    f"{new_default_version[sym]} (was {old_default_version[sym]}); "
-                    f"siblings {', '.join(consumer_libs)} pick up the new version."
-                ),
-                old_value=old_default_version[sym],
-                new_value=new_default_version[sym],
-                affected_libraries=consumer_libs,
-            ),
-        )
-
-    return findings
-
-
-def _soname_skew_findings(
-    old_members: list[BundleMember],
-    new_members: list[BundleMember],
-    cohorts: list[str],
-) -> list[BundleFinding]:
-    """Pure cohort-skew logic over already-read bundle members.
-
-    SONAME skew is **only** evaluated within explicitly declared cohorts —
-    each entry of *cohorts* is a cohort-key prefix (e.g. ``"libfoo_"``)
-    naming a set of libraries the release engineer asserts are co-versioned.
-    Libraries that match no declared cohort are never compared, so a normal
-    release that bumps an independent ``libfoo.so.1 → libfoo.so.2`` while an
-    unrelated ``libbar.so.1`` stays put is never reported. With an empty
-    *cohorts* list this returns nothing: there is no implicit lockstep
-    invariant to infer from filenames alone.
-    """
-    # An empty prefix (e.g. --bundle-cohort "" from an unset shell var) would
-    # be treated as "no filter" by the detector and compare every DSO —
-    # reintroducing the global false positive the opt-in exists to prevent.
-    # Strip and drop blanks so only genuine cohort prefixes are honoured.
-    prefixes = [p.strip() for p in cohorts if p and p.strip()]
-    if not prefixes:
-        return []
-    from .diff_cpp_patterns import detect_bundle_soname_skew
-
-    findings: list[BundleFinding] = []
-    for prefix in prefixes:
-        for change in detect_bundle_soname_skew(
-            old_members,
-            new_members,
-            cohort_prefix=prefix,
-        ):
-            findings.append(
-                BundleFinding(
-                    kind=change.kind,
-                    symbol=change.symbol,
-                    description=change.description,
-                    old_value=change.old_value,
-                    new_value=change.new_value,
-                    affected_libraries=list(change.affected_symbols or []),
-                )
-            )
-    return findings
-
-
-def _detect_soname_skew(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-    cohorts: list[str] | None,
-) -> list[BundleFinding]:
-    """Detect inconsistent SONAME major bumps within declared cohorts.
-
-    *cohorts* is the explicit opt-in: a list of cohort-key prefixes naming
-    co-versioned library sets (from ``compare-release --bundle-cohort``).
-    When it is empty/None nothing is emitted — there is no auto-grouping of
-    independent libraries by filename, which avoids false positives on
-    normal multi-library releases.
-
-    Members are derived from the *matched* release libraries
-    (``BundleSnapshot.libraries`` / ``.metadata``) rather than by rescanning
-    a single directory — release discovery is recursive, so a cohort member
-    living in another subdirectory must still participate. The authoritative
-    major comes from each library's DT_SONAME, falling back to the on-disk
-    filename; libraries with no derivable major (unversioned ``libfoo.so``)
-    are dropped.
-    """
-    cohorts = [c.strip() for c in (cohorts or []) if c and c.strip()]
-    if not cohorts:
-        return []
-    from .binary_utils import strip_vendor_hash
-    from .diff_cpp_patterns import BundleMember, _extract_soname_major
-
-    def _members(snap: BundleSnapshot) -> list[BundleMember]:
-        from .bundle_soname import resolved_basename
-
-        members: list[BundleMember] = []
-        for name, path in snap.libraries.items():
-            meta = snap.metadata.get(name)
-            # Resolved target's basename (matches capture_bundle_facts()'s
-            # identical use) -- but only when filesystem_backed: a facts-
-            # reconstructed snapshot's paths are synthetic (bare
-            # `Path("libfoo.so.1")`), and Path.resolve() would still walk
-            # CWD for those, letting an unrelated same-named CWD entry
-            # override the persisted basename (Codex review). Use it as-is.
-            real_name = resolved_basename(path) if snap.filesystem_backed else path.name
-            soname = (meta.soname if meta and meta.soname else "") or real_name
-            # G9 (remaining half): DT_SONAME is read directly off the ELF
-            # here, unlike `.library` (the on-disk filename), which the
-            # cohort key normalizes via strip_vendor_hash downstream. An
-            # auditwheel/delocate-vendored library's SONAME carries the same
-            # content-hash suffix as its filename (e.g.
-            # `libfoo_core-a1b2c3d4.so.2`) — strip it first so both the major
-            # extraction below and `BundleMember.soname` see the canonical
-            # logical SONAME, matching `.library`'s normalization. A hashed
-            # *versioned* dylib (`libfoo.2-a1b2c3.dylib`) has the hash
-            # between the major and the extension, so extracting the major
-            # from the raw string misses it entirely (Codex review).
-            stripped_soname = strip_vendor_hash(soname)
-            major = _extract_soname_major(stripped_soname)
-            if major is None:
-                major = _extract_soname_major(strip_vendor_hash(real_name))
-            if major is None:
-                continue
-            members.append(
-                BundleMember(
-                    library=real_name,
-                    soname=stripped_soname,
-                    soname_major=major,
-                )
-            )
-        return members
-
-    old_members = _members(old)
-    new_members = _members(new)
-    if not old_members or not new_members:
-        return []
-    return _soname_skew_findings(old_members, new_members, cohorts)
-
-
-def _entry_targets(entry: ManifestEntry) -> list[tuple[str, str]]:
-    """Decompose a manifest entry into ``[(display_name, match_kind)]``.
-
-    Where ``match_kind`` is one of:
-        - ``"symbol"`` — literal equality against ``.dynsym``.
-        - ``"pattern"`` — fnmatch glob against the *demangled* form.
-        - ``"template"`` — substring match against the demangled form.
-
-    A symbol or pattern entry yields one target; a template entry
-    yields **one target per instantiation**, so each instantiation is
-    matched (and reported) independently. The reviewer's regression:
-    a single template entry with four instantiations where only two
-    are exported previously short-circuited at "any match found" and
-    declared the entry satisfied. Per-instantiation decomposition
-    makes the contract explicit and gives users one finding per
-    missing instantiation.
-    """
-    if entry.symbol is not None:
-        return [(entry.symbol, "symbol")]
-    if entry.pattern is not None:
-        return [(entry.pattern, "pattern")]
-    expanded = _expand_instantiations(entry.template or "", entry.instantiations)
-    return [(t, "template") for t in expanded]
-
-
-def _build_demangled_index(snapshot: BundleSnapshot) -> list[tuple[str, str]]:
-    """Return ``[(demangled_name, library_name)]`` for every public export.
-
-    Performed once per :func:`_match_entry` call so manifest checking is
-    O(symbols + targets × index) rather than O(symbols × targets) — for
-    a large bundle (~50k exported symbols) with a manifest
-    containing hundreds of template instantiations, the naïve
-    re-scan-per-target path would dominate ``compare-release`` runtime
-    now that bundle analysis is default-on.
-
-    Demangling uses :func:`abicheck.demangle.demangle`; when the
-    demangler is unavailable, the mangled name is recorded so
-    ``extern "C"`` symbols still match.
-    """
-    from .demangle import demangle as _demangle
-
-    index: list[tuple[str, str]] = []
-    for lib_name, meta in snapshot.metadata.items():
-        for sym in meta.symbols:
-            if sym.visibility not in ("default", "protected"):
-                continue
-            index.append((_demangle(sym.name) or sym.name, lib_name))
-    return index
-
-
-def _match_target_against_index(
-    target: str,
-    kind: str,
-    snapshot: BundleSnapshot,
-    index: list[tuple[str, str]] | None = None,
-) -> tuple[list[str], list[ProviderEntry]]:
-    """Find every export in *snapshot* that satisfies *target* of *kind*.
-
-    Returns ``(matched_demangled_names, providers)``.  The provider list
-    has one :class:`ProviderEntry` per library that exports a matching
-    symbol (de-duplicated; one entry per library, not per symbol).
-
-    When *index* is supplied (a pre-built demangled-name → library
-    mapping), the scan operates against the cached list. Callers
-    iterating many targets against the same snapshot should pass a
-    shared index to amortise the demangle pass.
-    """
-    import fnmatch
-
-    if kind == "symbol":
-        providers = snapshot.resolution.providers_for(target)
-        return ([target] if providers else []), providers
-
-    if index is None:
-        index = _build_demangled_index(snapshot)
-
-    matched: list[str] = []
-    provider_set: set[str] = set()
-    for demangled, lib_name in index:
-        if lib_name in provider_set:
-            # We already recorded this library as a provider — one
-            # match per library is enough; skip the rest of its exports.
-            # (Avoids quadratic work when a library exports thousands
-            # of symbols matching a coarse pattern.)
-            continue
-        hit = False
-        if kind == "pattern":
-            hit = fnmatch.fnmatchcase(demangled, target)
-        else:  # template
-            hit = target in demangled
-        if hit:
-            matched.append(demangled)
-            provider_set.add(lib_name)
-    providers = [
-        ProviderEntry(library=name, version="") for name in sorted(provider_set)
-    ]
-    return matched, providers
-
-
-# Backward-compatibility alias for the original name — some tests and
-# external integrations imported _match_target directly. The new code
-# path is :func:`_match_target_against_index`.
-_match_target = _match_target_against_index
-
-
-def _match_entry(
-    entry: ManifestEntry,
-    snapshot: BundleSnapshot,
-    index: list[tuple[str, str]] | None = None,
-) -> list[tuple[str, str, list[str], list[ProviderEntry]]]:
-    """Return per-target match results for *entry*.
-
-    ``[(target_display_name, kind, matched_demangled, providers), ...]``
-
-    For ``symbol`` and ``pattern`` entries the list has one element.
-    For ``template`` entries it has one element per instantiation, so
-    a partially-satisfied template fires one ``MANIFEST_INSTANTIATION_REMOVED``
-    per missing instantiation rather than silently passing because some
-    sibling instantiation happened to match.
-
-    When the caller has many manifest entries to evaluate against the
-    same snapshot, build a shared index once via
-    :func:`_build_demangled_index` and pass it in to amortise the
-    O(symbols) demangle pass across all targets.
-    """
-    needs_index = any(kind != "symbol" for _, kind in _entry_targets(entry))
-    if index is None and needs_index:
-        index = _build_demangled_index(snapshot)
-    out: list[tuple[str, str, list[str], list[ProviderEntry]]] = []
-    for target, kind in _entry_targets(entry):
-        matched, providers = _match_target_against_index(target, kind, snapshot, index)
-        out.append((target, kind, matched, providers))
-    return out
-
-
-def _detect_manifest_drift(
-    old: BundleSnapshot,
-    new: BundleSnapshot,
-    manifest: InstantiationManifest,
-) -> list[BundleFinding]:
-    """Enforce a release manifest against the new bundle.
-
-    Decomposes template entries into one virtual target per
-    instantiation so each instantiation is checked independently.
-    Per-snapshot demangle indexes are built once and reused across
-    every manifest entry — manifest enforcement scales O(symbols +
-    Σtargets) rather than O(symbols × Σtargets).
-
-    For each target:
-      - If no exported symbol matches → ``BUNDLE_MANIFEST_INSTANTIATION_REMOVED``.
-      - If matched but at the wrong provider (when ``optional_provider=False``)
-        → ``BUNDLE_MANIFEST_INSTANTIATION_REMOVED`` (contract names the lib).
-
-    Symbols in the new bundle but not in the manifest are not flagged
-    here (out-of-manifest exports are not necessarily promised).
-    """
-    findings: list[BundleFinding] = []
-    # Build the per-snapshot demangle indexes once; both the
-    # "missing in new" and "newly promised" passes reuse them.
-    new_index = _build_demangled_index(new)
-    old_index = _build_demangled_index(old)
-
-    for entry in manifest.entries:
-        for target, kind, matched, providers in _match_entry(entry, new, new_index):
-            if not matched:
-                findings.append(
-                    BundleFinding(
-                        kind=ChangeKind.BUNDLE_MANIFEST_INSTANTIATION_REMOVED,
-                        symbol=target,
-                        description=(
-                            f"Manifest promises {kind} {target!r} but no "
-                            f"exported symbol in the new bundle matches it."
-                        ),
-                        provider_library=entry.library,
-                    ),
-                )
-                continue
-            if not entry.optional_provider and entry.library is not None:
-
-                def _matches(
-                    prov: ProviderEntry, _entry: ManifestEntry = entry
-                ) -> bool:
-                    if prov.library == _entry.library:
-                        return True
-                    meta = new.metadata.get(prov.library)
-                    return meta is not None and meta.soname == _entry.library
-
-                if not any(_matches(p) for p in providers):
-                    got = ", ".join(sorted(p.library for p in providers))
-                    findings.append(
-                        BundleFinding(
-                            kind=ChangeKind.BUNDLE_MANIFEST_INSTANTIATION_REMOVED,
-                            symbol=target,
-                            description=(
-                                f"Manifest requires {kind} {target!r} to be "
-                                f"provided by {entry.library}, but it is "
-                                f"provided by {got} instead."
-                            ),
-                            provider_library=entry.library,
-                            new_value=got,
-                        ),
-                    )
-
-    # Newly-promised targets — matched in new bundle but not in old.
-    for entry in manifest.entries:
-        new_targets = _match_entry(entry, new, new_index)
-        old_targets = {t: m for t, _, m, _ in _match_entry(entry, old, old_index)}
-        for target, kind, matched_new, _ in new_targets:
-            if not matched_new:
-                continue
-            if old_targets.get(target):
-                continue
-            findings.append(
-                BundleFinding(
-                    kind=ChangeKind.BUNDLE_MANIFEST_INSTANTIATION_ADDED,
-                    symbol=target,
-                    description=(
-                        f"Manifest now promises {kind} {target!r}; "
-                        f"not exported by the old bundle. New public surface."
-                    ),
-                    provider_library=entry.library,
-                ),
-            )
-
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Internal heuristics
-# ---------------------------------------------------------------------------
-
-# Common system-provided symbols imported by almost every C/C++ DSO.
-# Avoids false-positive bundle findings for libc/libstdc++ symbols.
-DEFAULT_SYSTEM_SYMBOLS: frozenset[str] = frozenset(
-    {
-        "__libc_start_main",
-        "__cxa_atexit",
-        "__cxa_finalize",
-        "__cxa_throw",
-        "__gxx_personality_v0",
-        "__stack_chk_fail",
-        "__stack_chk_guard",
-        "__tls_get_addr",
-        "__errno_location",
-        "_ITM_registerTMCloneTable",
-        "_ITM_deregisterTMCloneTable",
-        "abort",
-        "exit",
-        "malloc",
-        "free",
-        "calloc",
-        "realloc",
-        "memcpy",
-        "memmove",
-        "memset",
-        "memcmp",
-        "strlen",
-        "strcmp",
-        "strncmp",
-        "strcpy",
-        "strncpy",
-        "strdup",
-        "fprintf",
-        "printf",
-        "puts",
-        "pthread_once",
-        "pthread_self",
-        "pthread_create",
-        "pthread_join",
-    }
-)
-
-
-def _looks_system(soname: str) -> bool:
-    """Heuristic: looks like a system-provided library by name."""
-    return (
-        soname.startswith("libc.so")
-        or soname.startswith("libm.so")
-        or soname.startswith("libdl.so")
-        or soname.startswith("libpthread.so")
-        or soname.startswith("librt.so")
-        or soname.startswith("libstdc++.so")
-        or soname.startswith("libc++.so")
-        or soname.startswith("libgcc")
-        or soname.startswith("ld-linux")
-    )
-
-
-def _looks_system_symbol(name: str) -> bool:
-    """Heuristic: imported symbol that is almost certainly system-provided."""
-    if name.startswith("_ZNSt") or name.startswith("_ZSt"):
-        return True  # std:: mangled
-    if name.startswith("_ZNK") and "St" in name[:8]:
-        return True
-    return False
-
-
-# Symbol-version namespaces owned by the C/C++ runtime and toolchain. A symbol
-# imported with one of these required versions is satisfied by libc /
-# libstdc++ / libgcc / libgomp — never by a sibling inside the analysed
-# bundle — so it must not be reported as a dropped intra-bundle dependency.
-# This is the version-evidence half of the field-derived oneDAL fix
-# (``syscall@GLIBC_*``, ``stdout@GLIBC_*``, ``_ZdlPvm@CXXABI_*``).
-_SYSTEM_VERSION_PREFIXES: tuple[str, ...] = (
-    "GLIBC_",
-    "GLIBCXX_",
-    "CXXABI_",
-    "GCC_",
-    "LIBGCC_",
-    "LIBC_",
-    "GOMP_",
-    "OMP_",
-    "GFORTRAN_",
-    "GLIBCABI_",
-)
-
-
-def _looks_system_version(version: str) -> bool:
-    """True when a required symbol version is a C/C++ runtime/toolchain namespace."""
-    return any(version.startswith(prefix) for prefix in _SYSTEM_VERSION_PREFIXES)
-
-
-def _import_is_external(
-    consumer: ConsumerEntry,
-    consumer_meta: ElfMetadata,
-    snapshot: BundleSnapshot,
-) -> bool:
-    """Classify an import as external using version + provider evidence.
-
-    An import is external (never a *dropped intra-bundle* dependency) when
-    its required symbol version is satisfied by a library outside the bundle.
-
-    The precise signal is the **per-symbol verneed provider**
-    (``ConsumerEntry.version_soname``): GNU version labels are scoped per
-    verneed provider, not globally unique, so two providers can both advertise
-    e.g. ``FOO_1.0``. When that soname is known we resolve this exact import —
-    external iff its provider soname does not resolve inside the bundle —
-    without guessing from the bare label.
-
-    ``is_intra_bundle_provider`` matches by exact soname *and* filename stem, so
-    a SONAME-major bump (``libcore.so.1`` → ``libcore.so.2``) where a sibling
-    still NEEDs the old soname is still recognised as intra-bundle.
-
-    When the per-symbol verneed soname is unavailable, fall back to the
-    label-level scan over ``versions_required`` — provider evidence still wins
-    over the ``_looks_system_version`` toolchain-namespace heuristic.
-
-    Unversioned imports (``version == ""``) return ``False`` so an unversioned
-    internal sibling import still produces a finding: classify by
-    provider/version evidence, not only by symbol-name allow-lists.
-    """
-    version = consumer.version
-    if not version:
-        return False
-
-    def _resolves_intra(soname: str) -> bool:
-        # Two independent, complementary signals, either sufficient:
-        # resolution.soname_to_name is the exact map
-        # _reachable_intra_libraries()/the version_soname pinning above use
-        # (knows a member's real on-disk filename when retained through a
-        # differently named symlink alias with no DT_SONAME -- Codex
-        # review); is_intra_bundle_provider() additionally does filename-
-        # stem matching (a SONAME-major bump, e.g. libcore.so.1 ->
-        # libcore.so.2, where a sibling's DT_NEEDED still cites the old
-        # soname -- test_versioned_import_after_soname_bump_still_fires),
-        # which the exact map alone does not cover since it only indexes a
-        # library's *current* soname/canonical key/real filename, never a
-        # soname it used to advertise.
-        return (
-            soname in snapshot.resolution.soname_to_name
-            or snapshot.is_intra_bundle_provider(soname)
-        )
-
-    # Preferred path: the exact verneed provider for *this* symbol is known.
-    # No label-collision ambiguity — resolve this import directly.
-    if consumer.version_soname:
-        return not _resolves_intra(consumer.version_soname)
-    # Fallback: per-symbol verneed soname not captured. Scan the label across
-    # all verneed providers. If any soname carrying this label resolves inside
-    # the bundle, treat the import as intra (keep the finding); only when every
-    # provider of the label is external is it external. Provider evidence wins
-    # over the system-namespace shortcut so a vendored runtime stays visible.
-    external_match = False
-    for soname, versions in consumer_meta.versions_required.items():
-        if version not in versions:
-            continue
-        if _resolves_intra(soname):
-            return False  # required from a bundle sibling — keep the finding
-        external_match = True
-    if external_match:
-        return True
-    return _looks_system_version(version)
-
-
-_ELF_MAGIC = b"\x7fELF"
-
-
-def _path_looks_like_elf(path: Path) -> bool:
-    """Cheap ELF-magic sniff. Avoids spurious warnings from
-    :func:`parse_elf_metadata` on JSON snapshot inputs and other non-ELF
-    artefacts present in a release directory."""
-    try:
-        with open(path, "rb") as f:
-            return f.read(4) == _ELF_MAGIC
-    except OSError:
-        return False
-
-
-def _strip_namespace_prefix(name: str) -> str:
-    """Return the unqualified component of a possibly C++-qualified name.
-
-    Used by :func:`_detect_intra_type_changed` to find type references
-    inside mangled symbols even when the diff reports the type by its
-    fully-qualified name.
-    """
-    if "::" in name:
-        return name.rsplit("::", 1)[-1]
-    return name
