+++ abicheck/cli_options.py
@@ -33,14 +33,10 @@ import click
 from .cli_params import (
     BUILTIN_POLICY_PROFILES,
     DEFAULT_POLICY_PROFILE,
-    DEPTH_PARAM,
     POLICY_FILE_PARAM,
-    SIDED_BUILD_INFO_PARAM,
     SIDED_DUMP_MANIFEST_PARAM,
-    SIDED_EXISTING_PATH_PARAM,
     SIDED_INCLUDE_PATH_PARAM,
     SIDED_PATH_PARAM,
-    SIDED_SOURCES_PARAM,
     SIDED_STR_PARAM,
     SidedChoiceParam,
 )
@@ -801,11 +797,14 @@ def compile_context_options(*, sided_frontend: bool = False) -> Callable[[F], F]
             "--ast-frontend",
             "header_backend",
             **frontend_kwargs,
-            help=("Scope to one side with an 'old='/'new=' prefix, repeating the "
-            "flag per side (e.g. --ast-frontend old=castxml --ast-frontend "
-            "new=clang) when the old release parses on one frontend and the new "
-            "one needs the other; a bare value applies to both (default: auto). "
-            if sided_frontend else "")
+            help=(
+                "Scope to one side with an 'old='/'new=' prefix, repeating the "
+                "flag per side (e.g. --ast-frontend old=castxml --ast-frontend "
+                "new=clang) when the old release parses on one frontend and the new "
+                "one needs the other; a bare value applies to both (default: auto). "
+                if sided_frontend
+                else ""
+            )
             + "C/C++ AST frontend (ADR-037 D8): castxml (default schema reference) "
             "or clang (-ast-dump=json; for hosts where castxml is absent or its "
             "bundled frontend chokes). hybrid (G28 Phase 3) runs BOTH and merges "
@@ -1340,496 +1339,13 @@ def artifact_set_options(func: F) -> F:
     return func
 
 
-def release_options(func: F) -> F:
-    """Directory/package (release) comparison knobs, folded onto ``compare``.
-
-    The release-only options the removed ``compare-release`` command exposed:
-    package extraction (``--debug-info*``/``--devel-pkg*``), DSO selection
-    (``--include-private-dso``/``--keep-extracted``), the removed-library gate, and
-    the ADR-023 bundle/manifest analysis. They bite only when ``compare``'s
-    operands are directories or packages (the per-library fan-out); on single-file
-    inputs they are inert. Declared once here so ``compare`` and the internal
-    release engine share one surface (ADR-037 D7). Applied bottom-up, so listed in
-    reverse of displayed order.
-    """
-    func = click.option(
-        "--no-bundle-analysis",
-        "no_bundle_analysis",
-        is_flag=True,
-        default=False,
-        help="Skip bundle-level cross-library analysis (debug/parity escape hatch). "
-        "Bundle findings catch intra-bundle symbol removals, signature drift "
-        "across DSO boundaries, type drift across siblings, provider migration, "
-        "and manifest mismatches. (directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--bundle-cohort",
-        "bundle_cohorts",
-        multiple=True,
-        metavar="PREFIX",
-        help="Declare a co-versioned library cohort by name prefix (e.g. "
-        "'libfoo_'). Repeatable. Enables the BUNDLE_SONAME_SKEW check. "
-        "(directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--bundle-system-providers",
-        "bundle_system_providers",
-        default="",
-        help="Comma-separated extra sonames to treat as system-provided "
-        "(extends the built-in libc/libstdc++/libgcc/libtbb allow-list). "
-        "(directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--manifest",
-        "manifest_path",
-        type=click.Path(exists=True, path_type=Path),
-        default=None,
-        help="ABI instantiation manifest (YAML/JSON) listing symbols the release "
-        "publicly promises (ADR-023). (directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--bundle-facts-out",
-        "bundle_facts_out",
-        type=click.Path(path_type=Path),
-        default=None,
-        help="Persist this run's OLD-side bundle facts (per-library snapshots "
-        "plus the instantiation manifest, if any) to PATH (G38 Phase 2, "
-        "ADR-023 amendment) for a later stored-baseline bundle comparison. "
-        "Additive output alongside the ordinary live-vs-live comparison; "
-        "no-op with --no-bundle-analysis. (directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--keep-extracted",
-        "keep_extracted",
-        is_flag=True,
-        default=False,
-        help="Keep extracted temporary files for debugging. "
-        "(directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--include-private-dso",
-        "include_private_dso",
-        is_flag=True,
-        default=False,
-        help="Include private (non-public) shared objects from non-standard "
-        "paths. (directory/package inputs only)",
-    )(func)
-    func = click.option(
-        "--devel-pkg",
-        "devel_pkg",
-        multiple=True,
-        type=SIDED_EXISTING_PATH_PARAM,
-        help="Development package with headers, scoped per side with an "
-        "'old='/'new=' prefix (e.g. --devel-pkg old=a-dev.rpm --devel-pkg "
-        "new=b-dev.rpm). Directory/package inputs only (ADR-040).",
-    )(func)
-    func = click.option(
-        "--debug-info",
-        "debug_info",
-        multiple=True,
-        type=SIDED_EXISTING_PATH_PARAM,
-        help="Debug info package (RPM/Deb/tar), scoped per side with an "
-        "'old='/'new=' prefix (e.g. --debug-info old=a-dbg.rpm --debug-info "
-        "new=b-dbg.rpm). Directory/package inputs only (ADR-040).",
-    )(func)
-    func = click.option(
-        "--fail-on-removed-library/--no-fail-on-removed-library",
-        "fail_on_removed",
-        default=False,
-        help="Exit 8 when a library present in old_dir is absent in new_dir. "
-        "(directory/package inputs only)",
-    )(func)
-    return func
-
-
-def debug_resolution_options(func: F) -> F:
-    """Separate-debug-file resolution (ADR-021a): roots + debuginfod + format.
-
-    Currently a ``compare``-only family — it resolves *local* ELF debug
-    artifacts, which the package-oriented (``compare-release``) and
-    snapshot-oriented (``appcompat``) commands do not take. It
-    lives here so the moment a second command needs it there is one definition to
-    compose, not a copy to drift (ADR-037 D3).
-    """
-    func = click.option(
-        "--dwarf",
-        "debug_format",
-        flag_value="dwarf",
-        hidden=True,
-        help="Force DWARF debug format for both sides (ELF only).",
-    )(func)
-    func = click.option(
-        "--ctf",
-        "debug_format",
-        flag_value="ctf",
-        hidden=True,
-        help="Force CTF debug format for both sides (ELF only).",
-    )(func)
-    func = click.option(
-        "--btf",
-        "debug_format",
-        flag_value="btf",
-        default=None,
-        hidden=True,
-        help="Force BTF debug format for both sides (ELF only).",
-    )(func)
-    func = click.option(
-        "--debug-format",
-        "debug_format_opt",
-        type=click.Choice(["auto", "dwarf", "btf", "ctf"], case_sensitive=False),
-        default=None,
-        hidden=True,
-        help="Force the ELF debug format for both sides (auto=pick best available). "
-        "Supersedes the individual --btf/--ctf/--dwarf flags. Demoted to the "
-        "debug.format config key (ADR-040 L2); this flag still overrides it.",
-    )(func)
-    func = click.option(
-        "--debuginfod-url",
-        "debuginfod_url",
-        default=None,
-        hidden=True,
-        help="debuginfod server URL (overrides DEBUGINFOD_URLS env var). Demoted to "
-        "the debug.debuginfod_url config key (ADR-040 L2); this flag still overrides it.",
-    )(func)
-    func = click.option(
-        "--debuginfod/--no-debuginfod",
-        "debuginfod",
-        default=False,
-        hidden=True,
-        help="Enable debuginfod network resolution for debug info (opt-in). Demoted "
-        "to the debug.debuginfod config key (ADR-040 L2); --debuginfod/--no-debuginfod "
-        "still overrides it either way.",
-    )(func)
-    func = click.option(
-        "--debug-root",
-        "debug_root",
-        multiple=True,
-        type=SIDED_PATH_PARAM,
-        help="Directory containing separate debug files (build-id trees, "
-        "path-mirror, dSYM bundles). Applies to both sides; scope to one with an "
-        "'old='/'new=' prefix, repeating the flag per side "
-        "(e.g. --debug-root old=dbg1 --debug-root new=dbg2). Repeatable (ADR-040).",
-    )(func)
-    func = click.option(
-        "--dwarf-only/--no-dwarf-only",
-        "dwarf_only",
-        default=False,
-        hidden=True,
-        help="Force DWARF-only mode for both sides: use DWARF debug info "
-        "as primary data source even when headers are available. Demoted to the "
-        "debug.dwarf_only config key (ADR-040 L2); --dwarf-only/--no-dwarf-only "
-        "still overrides it either way (e.g. --no-dwarf-only restores header parsing "
-        "for a one-off run).",
-    )(func)
-    return func
-
-
-def adr027_compare_options(func: F) -> F:
-    """Add the ADR-027 API-surface-intelligence options to ``compare``.
-
-    ``--pattern-verdicts`` / ``--explain-patterns`` (A4 modulation) and
-    ``--surface-metrics`` (A1/D1.2 metric drift). Decorators apply bottom-up, so
-    they are listed here in reverse of their displayed order.
-    """
-    func = click.option(
-        "--surface-metrics",
-        "surface_metrics",
-        is_flag=True,
-        default=False,
-        help="Emit aggregate public-surface metric drift (ADR-027): "
-        "public_surface_grew/shrank, undocumented_export_ratio_increased. "
-        "Informational (COMPATIBLE).",
-    )(func)
-    func = click.option(
-        "--explain-patterns",
-        "explain_patterns",
-        is_flag=True,
-        default=False,
-        help="Print idiom evidence behind each modulation (implies "
-        "--pattern-verdicts).",
-    )(func)
-    func = click.option(
-        "--pattern-verdicts/--no-pattern-verdicts",
-        "pattern_verdicts",
-        default=False,
-        help="Modulate verdicts with idiom/anti-pattern evidence (ADR-027): "
-        "demote opaque-pointer/PIMPL-hidden layout changes (header-aware only) "
-        "and raise breaks when an opacity/handle guarantee is lost. Disclosed in "
-        "the pattern_modulations ledger; reversible.",
-    )(func)
-    return func
-
-
-def app_usage_scope_options(func: F) -> F:
-    """Add the ADR-043 app-usage/required-symbol scoping options to ``compare``.
-
-    ``--used-by`` and ``--required-symbol``/
-    ``--required-symbols`` are mutually exclusive scoping mechanisms folding
-    the former standalone ``appcompat``/``plugin-check`` commands into
-    ``compare``. Decorators apply bottom-up, so they are listed here in
-    reverse of their displayed order.
-    """
-    func = click.option(
-        "--required-symbols",
-        "required_symbols_file",
-        type=click.Path(exists=True, dir_okay=False, path_type=Path),
-        default=None,
-        help="File of required symbols, one per line (blank lines and '#' "
-        "comments ignored). Combined with any --required-symbol values.",
-    )(func)
-    func = click.option(
-        "--required-symbol",
-        "required_symbols_opt",
-        multiple=True,
-        help="An exported linker symbol a plugin host resolves via dlopen/dlsym "
-        "and requires (repeatable; folds `plugin-check`). Scopes the "
-        "comparison to this explicit entrypoint contract instead of the "
-        "full diff. Mutually exclusive with --used-by.",
-    )(func)
-    func = click.option(
-        "--used-by",
-        "used_by_apps",
-        multiple=True,
-        type=click.Path(exists=True, dir_okay=False, path_type=Path),
-        help="Application binary whose actual imports/required symbol versions "
-        "scope the comparison (repeatable; folds `appcompat`). The full "
-        "library comparison still runs once; the worst app-scoped result "
-        "becomes the primary verdict/exit code, with the full verdict and "
-        "unrelated changes kept as informational context. OLD/NEW may be "
-        "real library binaries or JSON snapshots carrying binary evidence "
-        "(a `dump` of a real library, not headers-only). Mutually "
-        "exclusive with --required-symbol/--required-symbols.",
-    )(func)
-    return func
-
-
-def build_source_dump_options(func: F) -> F:
-    """Add the ``--build-info`` / ``--sources`` embed options to ``dump``.
-
-    Source-tree-centric inputs (ADR-028..033 amendment): ``--sources`` is a
-    source checkout — L4 source ABI replay and the L5 graph are run inline and
-    embedded; ``--build-info`` is an optional build dir / ``compile_commands.json``
-    / pre-built pack supplying L3 (auto-discovered inside the source tree when
-    omitted). Either flag also accepts, and auto-detects, a build-emitted
-    ``abicheck_inputs/`` Flow-2 pack directory or a pre-built ``BuildSourcePack``
-    directory (from an internal/producer-side collection step) — both are
-    ingested and validated automatically, no separate ``inputs validate``/
-    ``merge`` step needed (ADR-043 D1). Embedding makes the ``.abi.json``
-    self-contained, so a later ``compare old.json new.json`` carries the facts
-    with no out-of-band directories. Applied bottom-up, so listed in reverse of
-    display.
-    """
-    func = click.option(
-        "--depth",
-        "depth",
-        type=DEPTH_PARAM,
-        default=None,
-        help="Evidence-depth dial (same vocabulary as `compare`/`scan --depth`): "
-        "binary=symbols only, headers=+header AST (default), build=+build "
-        "context, source=+source replay & call graph.",
-    )(func)
-    func = click.option(
-        "--allow-build-query",
-        "allow_build_query",
-        is_flag=True,
-        default=False,
-        hidden=True,  # deprecated no-op (ADR-032 amended): build query is now automatic
-        help="Deprecated and ignored. Build-system queries now run automatically "
-        "when --sources is given (abicheck infers and runs cmake/make/bazel "
-        "itself); no flag is needed. Kept as a no-op for backward compatibility.",
-    )(func)
-    func = click.option(
-        "--config",
-        "build_config",
-        type=click.Path(exists=True, dir_okay=False, path_type=Path),
-        default=None,
-        help="Path to the project `.abicheck.yml` (ADR-037 D4): build system, "
-        "query command, compile-DB location, plus the stable severity/scope/"
-        "suppression/source settings. Defaults to `.abicheck.yml` at the "
-        "--sources tree root for non-executing settings; build.query runs only "
-        "from an explicit --config.",
-    )(func)
-    func = click.option(
-        "--build-compile-db",
-        "build_compile_db",
-        default=None,
-        metavar="GLOB",
-        help="Where a build/query lands its compile_commands.json, relative to "
-        "--sources (e.g. 'build/compile_commands.json'). CLI equivalent of "
-        "`.abicheck.yml` build.compile_db; overrides it when both are given.",
-    )(func)
-    func = click.option(
-        "--build-query",
-        "build_query",
-        default=None,
-        metavar="CMD",
-        help="Override the inferred build-system query command that emits a "
-        "compile DB without a full build (e.g. 'cmake -S . -B build "
-        "-DCMAKE_EXPORT_COMPILE_COMMANDS=ON'). CLI equivalent of `.abicheck.yml` "
-        "build.query — runs automatically as trusted operator input. Usually "
-        "unnecessary: with just --sources, abicheck infers and runs the query "
-        "itself.",
-    )(func)
-    func = click.option(
-        "--build-target",
-        "build_targets",
-        multiple=True,
-        metavar="TARGET",
-        help="Explicit build-system root target(s) to scope L3 evidence "
-        "collection to, instead of a workspace-wide query (P0.2; Bazel "
-        "only so far, e.g. '//:math'). Repeatable — each root's transitive "
-        "dependency closure is unioned. CLI equivalent of `.abicheck.yml` "
-        "build.targets; overrides it when both are given. Without this, a "
-        "multi-package workspace with fixture/test targets alongside the "
-        "real library is collected in full, which can pollute L3 evidence "
-        "with unrelated compile units.",
-    )(func)
-    func = click.option(
-        "--sources",
-        "sources",
-        type=click.Path(exists=True, path_type=Path),
-        default=None,
-        help="Source checkout to run source-ABI replay and build the call "
-        "graph over, embedding both inline. (An existing pack directory — e.g. "
-        "from the abicheck-cc wrapper or Clang plugin — is auto-detected by "
-        "its manifest.json and loaded as that pack instead.)",
-    )(func)
-    func = click.option(
-        "--build-info",
-        "build_info",
-        type=click.Path(exists=True, path_type=Path),
-        default=None,
-        help="Optional build context: a build dir, a compile_commands.json, "
-        "or a pre-captured pack. Auto-discovered inside the --sources tree when "
-        "omitted. When it resolves to a compile database and -H/--header is "
-        "given, that database also parameterizes the header parse with the "
-        "build's exact flags (scope it with --compile-db-filter).",
-    )(func)
-    return func
-
-
-def header_graph_options(func: F) -> F:
-    """The shared, deprecated ``--header-graph``/``--header-graph-includes`` pair.
-
-    G29 Phase A: the L2 header-only semantic graph
-    (:func:`~abicheck.buildsource.header_graph.build_header_only_graph`) — and
-    its include-file extension — is now always built whenever headers are
-    available (``--depth headers`` or deeper), for both ``compare`` and
-    ``dump``. These two flags are no longer opt-in toggles; they are kept as
-    *hidden*, inert no-op shims (``hidden=True`` — absent from ``--help`` and
-    from ``tests/test_cli_contract.py``'s ``_OPTION_SET_SNAPSHOT``) purely so
-    an existing script/CI invocation that still passes ``--header-graph``
-    doesn't hard-fail with "no such option". Passing either flag prints a
-    one-line deprecation note to stderr and otherwise changes nothing — the
-    graph is built identically whether or not the flag is given. Planned
-    removal: two minor releases after this change ships (track in
-    CHANGELOG.md). Shared by ``compare`` and ``dump`` so the two flags' spelling
-    can never drift between them. Applied bottom-up, so listed in reverse of
-    display.
-    """
-    func = click.option(
-        "--header-graph-includes",
-        "header_graph_includes_deprecated",
-        is_flag=True,
-        default=False,
-        hidden=True,
-        help="Deprecated, no-op: the include-file graph pass is now always run "
-        "alongside --header-graph's replacement (always-on L2 header graph). "
-        "Planned removal: two minor releases out.",
-    )(func)
-    func = click.option(
-        "--header-graph",
-        "header_graph_deprecated",
-        is_flag=True,
-        default=False,
-        hidden=True,
-        help="Deprecated, no-op: the L2 header-only semantic graph (ADR-041 "
-        "addendum) is now always built for --depth headers and above. Planned "
-        "removal: two minor releases out.",
-    )(func)
-    return func
-
-
-def warn_deprecated_header_graph_flags(
-    header_graph_deprecated: bool, header_graph_includes_deprecated: bool
-) -> None:
-    """Emit a deprecation note for the inert ``--header-graph``/``-includes`` shim.
-
-    Called from ``compare``/``dump_cmd`` bodies (not the Click callback
-    itself, so it runs after Click has finished parsing) whenever either
-    flag was passed on the command line. Behavior is identical either way —
-    this is purely a stderr note, per the "hidden shim must not control
-    behavior" policy (AGENTS.md deprecation convention).
-    """
-    if header_graph_deprecated or header_graph_includes_deprecated:
-        click.echo(
-            "Note: --header-graph/--header-graph-includes are deprecated "
-            "no-ops — the L2 header-only semantic graph is now always built "
-            "for --depth headers and above. Planned removal: two minor "
-            "releases out.",
-            err=True,
-        )
-
-
-def evidence_options(func: F) -> F:
-    """The shared two-sided evidence family (ADR-037 D3's ``@evidence_options``).
-
-    The single source of truth for the depth/source/build-info surface a
-    *two-sided* verdict command exposes: ``--depth`` plus the per-side
-    ``--old/new-sources`` and ``--old/new-build-info`` packs. ``dump`` is
-    single-sided (one artifact, plus the build-query knobs) so it composes the
-    sibling :func:`build_source_dump_options` instead — they are deliberately not
-    one decorator because their surfaces differ (per-side vs build-query), which
-    is why ``evidence`` is a registered-but-not-required family (only commands
-    that take source depth compose it).
-
-    By default ``compare old.json new.json`` reads build-info + source facts
-    **embedded** in each snapshot (single-artifact UX). The optional side-aware
-    ``--build-info`` and ``--sources`` (ADR-040) point at out-of-band pack
-    directories to supply or override those facts — for both sides, or per side
-    with an ``old=``/``new=`` prefix; ``--depth`` selects how deep the inline
-    collection runs (ADR-037 D5). All folded into the verdict as ordinary
-    findings, never overriding artifact-backed ABI verdicts (ADR-028 D3).
-    Applied bottom-up, so listed in reverse of displayed order.
-    """
-    func = click.option(
-        "--depth",
-        "depth",
-        type=DEPTH_PARAM,
-        default=None,
-        help="Evidence-depth dial: binary=symbols only, headers=+header AST "
-        "(default), build=+build context, source=+source replay & call graph. "
-        "Deeper-than-headers needs --sources or --build-info.",
-    )(func)
-    func = click.option(
-        "--sources",
-        "sources",
-        multiple=True,
-        type=SIDED_SOURCES_PARAM,
-        help="Source checkout for --depth build/source (collected inline, "
-        "embedding build/source/graph facts) or a pre-built `collect` pack, "
-        "overriding embedded. Applies to both sides; scope to one with an "
-        "'old='/'new=' prefix, repeating the flag per side "
-        "(e.g. --sources old=src_v1 --sources new=src_v2) (ADR-040).",
-    )(func)
-    func = click.option(
-        "--build-info",
-        "build_info",
-        multiple=True,
-        type=SIDED_BUILD_INFO_PARAM,
-        help="Out-of-band build context: a build dir, a compile_commands.json, "
-        "or a pack, overriding embedded. Applies to both sides; scope to one "
-        "with an 'old='/'new=' prefix, repeating the flag per side "
-        "(e.g. --build-info old=b1 --build-info new=b2) (ADR-040).",
-    )(func)
-    return func
-
-
-#: Back-compat alias for the pre-ADR-037-D3 name. ``evidence_options`` is the
-#: canonical spelling (the D3 table); this keeps existing imports working.
-build_source_compare_options = evidence_options
-
-
+#: ``compare``'s release-fanout/build-source/header-graph/evidence option
+#: groups moved to ``frontends/cli/options/release.py`` when this module
+#: reached the 2000-line hard cap -- the same split, for the same reason, as
+#: ``cli_profiles.py``/``cli_options_contract.py`` before it. Re-exported
+#: here (``X as X``, so the re-export is explicit to mypy) because every
+#: existing caller -- and every existing test importing them from here --
+#: reaches these decorators through this module.
 #: ADR-037 D10 CLI-contract metadata (family/flag tables, the compare
 #: flag-count budget ledger, and :func:`count_visible_options`) moved to
 #: ``cli_options_contract.py`` when this module reached the 2000-line hard
@@ -1869,6 +1385,17 @@ from .frontends.cli.options.profiles import (  # noqa: E402
     COMPARE_PROFILES as COMPARE_PROFILES,
     RUN_PROFILE_META_KEY as RUN_PROFILE_META_KEY,
     profile_option as profile_option,
+)
+from .frontends.cli.options.release import (  # noqa: E402
+    adr027_compare_options as adr027_compare_options,
+    app_usage_scope_options as app_usage_scope_options,
+    build_source_compare_options as build_source_compare_options,
+    build_source_dump_options as build_source_dump_options,
+    debug_resolution_options as debug_resolution_options,
+    evidence_options as evidence_options,
+    header_graph_options as header_graph_options,
+    release_options as release_options,
+    warn_deprecated_header_graph_flags as warn_deprecated_header_graph_flags,
 )
