+++ architecture/modules.yaml
@@ -268,6 +268,8 @@
         "abicheck/buildsource/baseline_publish.py",
         "abicheck/bundle.py",
         "abicheck/bundle_analysis.py",
+        "abicheck/bundle_detector_heuristics.py",
+        "abicheck/bundle_detectors.py",
         "abicheck/bundle_facts.py",
         "abicheck/bundle_facts_serialization.py",
         "abicheck/bundle_models.py",
@@ -363,6 +365,8 @@
         "abicheck/cli_compare_receipt.py",
         "abicheck/cli_compare_release.py",
         "abicheck/cli_compare_release_helpers.py",
+        "abicheck/cli_compare_release_matrix.py",
+        "abicheck/cli_compare_release_pairwise.py",
         "abicheck/cli_datasources.py",
         "abicheck/cli_dump_depth.py",
         "abicheck/cli_dump_dry_run_build_query.py",
@@ -417,6 +421,8 @@
       "cli_compare_receipt.py",
       "cli_compare_release.py",
       "cli_compare_release_helpers.py",
+      "cli_compare_release_matrix.py",
+      "cli_compare_release_pairwise.py",
       "cli_datasources.py",
       "cli_dump_depth.py",
       "cli_dump_dry_run_build_query.py",
@@ -505,6 +511,8 @@
     ],
     "bundle_": [
       "bundle_analysis.py",
+      "bundle_detector_heuristics.py",
+      "bundle_detectors.py",
       "bundle_facts.py",
       "bundle_facts_serialization.py",
       "bundle_manifest.py",
@@ -593,6 +601,8 @@
     "build_mode.py",
     "bundle.py",
     "bundle_analysis.py",
+    "bundle_detector_heuristics.py",
+    "bundle_detectors.py",
     "bundle_facts.py",
     "bundle_facts_serialization.py",
     "bundle_manifest.py",
@@ -624,6 +634,8 @@
     "cli_compare_receipt.py",
     "cli_compare_release.py",
     "cli_compare_release_helpers.py",
+    "cli_compare_release_matrix.py",
+    "cli_compare_release_pairwise.py",
     "cli_datasources.py",
     "cli_dump_depth.py",
     "cli_dump_dry_run_build_query.py",
