+++ tests/test_compare_release_contract_coverage.py
@@ -130,7 +130,7 @@ def test_compare_one_library_stamps_contract_coverage_failure_count(
         return CompareResult(diff=result, old_snapshot=old_snap, new_snapshot=new_snap)
 
     monkeypatch.setattr(
-        "abicheck.cli_compare_release._run_compare_pair", _fake_run_compare_pair
+        "abicheck.cli_compare_release_pairwise._run_compare_pair", _fake_run_compare_pair
     )
     entry = _compare_one_library(
         "libfoo.so",
@@ -184,7 +184,7 @@ def test_compare_one_library_stamps_coverage_warnings_onto_the_entry(
         return CompareResult(diff=result, old_snapshot=old_snap, new_snapshot=new_snap)
 
     monkeypatch.setattr(
-        "abicheck.cli_compare_release._run_compare_pair", _fake_run_compare_pair
+        "abicheck.cli_compare_release_pairwise._run_compare_pair", _fake_run_compare_pair
     )
     entry = _compare_one_library(
         "libfoo.so",
@@ -234,7 +234,7 @@ def test_compare_one_library_omits_coverage_warnings_key_when_none(
         return CompareResult(diff=result, old_snapshot=old_snap, new_snapshot=new_snap)
 
     monkeypatch.setattr(
-        "abicheck.cli_compare_release._run_compare_pair", _fake_run_compare_pair
+        "abicheck.cli_compare_release_pairwise._run_compare_pair", _fake_run_compare_pair
     )
     entry = _compare_one_library(
         "libfoo.so",
@@ -350,7 +350,7 @@ def test_compare_one_library_stashes_old_snapshot_only_when_requested(
         return CompareResult(diff=result, old_snapshot=old_snap, new_snapshot=new_snap)
 
     monkeypatch.setattr(
-        "abicheck.cli_compare_release._run_compare_pair", _fake_run_compare_pair
+        "abicheck.cli_compare_release_pairwise._run_compare_pair", _fake_run_compare_pair
     )
     common = (
         {"libfoo.so": old_path},
